apache2: fix the installed but not shipped issue
CQID: WIND00440334

We will get the following ERROR/WARN if we enable the
installed-vs-shipped check in QA:

ERROR: QA Issue: apache2: Files/directories were installed but not shipped
  /run
  /run/apache2

We already have the volatiles file to create them on startup, so remove
*/var/run* would fix the problem, note we remove /var/run but not /run
(and there is no /run in the image ($D) directory), the fixup_perms() function
in package.bbclass will create the /run in the package ($PKGD) directory when
there is /var/run (since /var/run symlinks to /run on the target), and then we
will get the warning, this is not a bug of fixup_perms(), but intended, see
package.bbclass for more details.

This only happens when --enable-layout=Debian which is enabled by us, so
it doesn't need to go upstream.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>