net-tools: fix do_patch_v2 breakage
JIRAID: CGP6-357

The net-tools can't be built if a previous build has failed, because
do_patch_v2 tries to re-run, and ends up re-applying the diff that
created the patches, which then results in them all trying to run
twice, and thus failing.

The net-tools maintained patches by quilt, while re-applying the
patches, it restored the old patches first and then apply the new
ones. When the user added a new patch to net-tools, the quilt rather
than bitbake's patch_do_patch should be used.

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>