grub-efi: override do_deploy() implemented in oe-core
The grub core image is created by grub-mkimage which is called twice.
One place is in this recipe's do_install() and another is in
do_deploy() in oe-core. Obviously, the creation of grub core image
occurring in do_deploy() is too late for signing. Thefore, override
do_deploy() implemented in oe-core, and leave the unsigned image in
${B} for manual signing. Also, make sure the image in ${DEPLOYDIR} is
signed if SCP is used.

Without this change, the grub core image in export/dist sometimes is
the unsigned one. This is unexpected result for SCP user.

Signed-off-by: Lans Zhang <jia.zhang@windriver.com>