grub-efi: fix error message during the boot
Issue: LIN8-4660

An error message printed by grub-efi flashes during the
boot. It is harmless but have a bad UX.

The cfg file provided by oe-core actually is used to set the
global variables root and prefix used to locate the position
of bootloader and grub.cfg, with the content as below:

search.file ($cmdpath)/EFI/BOOT/grub.cfg root
set prefix=($root)/EFI/BOOT

However, cmdpath is not defined at all, which caused the root
and prefix are wrongly defined. If this happens, grub-efi will
print the error message and re-calculate the values for them,
and this is why this error is harmless.

Signed-off-by: Lans Zhang <jia.zhang@windriver.com>